+++ sympy/physics/continuum_mechanics/__init__.py
@@ -1,0 +1,1 @@
+from .beam import Beam, PointLoad, DistributedLoad
