+++ sympy/physics/continuum_mechanics/beam.py
